refactor(webmcp): extract server-safe catalog
