+++ apps/web/app/verify/[hash]/VerifyPageWebMcpTools.tsx
@@ -7,9 +7,9 @@ import {
   type PublicVerificationRecord,
 } from "@/lib/verification/types";
 import {
-  CURRENT_VERIFICATION_HASH_HEX_LENGTH,
-  VERIFICATION_RECORD_TTL_DAYS,
-} from "@/lib/verification/constants";
+  VERIFICATION_EXPLANATION,
+  verificationCodeFormat,
+} from "@/lib/webmcp/catalog";
 import {
   WEBMCP_EMPTY_INPUT_SCHEMA,
   WEBMCP_READ_ONLY_ANNOTATIONS,
@@ -25,25 +25,6 @@ interface VerifyPageWebMcpToolsProps {
   record: PublicVerificationRecord;
 }
 
-// Source: the public `about.verification.*` copy and lib/verification/hmac.ts.
-// Keep the guarantees and limits aligned with that user-facing explanation.
-const VERIFICATION_EXPLANATION = {
-  algorithm: "HMAC-SHA256",
-  howItWorks:
-    `Current Chapa badges use a deterministic payload from the badge profile fields, sign it with a server-held secret key, and use the first ${CURRENT_VERIFICATION_HASH_HEX_LENGTH} hexadecimal characters (128 bits) as the verification code.`,
-  proves: [
-    "Only Chapa can issue the hash for the original signed payload because only the Chapa server knows the signing secret.",
-    "Changing any field in that original payload would produce a different hash.",
-    "The stored verification record exposes a subset of the original values for manual comparison and binds them to a specific date.",
-  ],
-  doesNotProve: [
-    "This lookup does not recompute the HMAC from an SVG, and the stored record does not expose every signed payload field for manual comparison.",
-    "It does not independently prove that the underlying platform data is accurate; Chapa trusts its platform data sources.",
-    "It does not prevent someone from editing an SVG file; it makes changes to signed fields detectable.",
-    `It is not a blockchain or permanent public ledger; verification records expire after ${VERIFICATION_RECORD_TTL_DAYS} days.`,
-  ],
-} as const;
-
 export function VerifyPageWebMcpTools({
   hash,
   record,
@@ -53,9 +34,7 @@ export function VerifyPageWebMcpTools({
     if (!webmcpEnabled) return [];
     const publicRecord = toPublicVerificationRecord(record);
 
-    const codeFormat = hash.length === CURRENT_VERIFICATION_HASH_HEX_LENGTH
-      ? `Current ${CURRENT_VERIFICATION_HASH_HEX_LENGTH}-character verification code.`
-      : `Verified legacy ${hash.length}-character verification code.`;
+    const codeFormat = verificationCodeFormat(hash);
 
     return [
       {
