+++ apps/web/lib/webmcp/errors.ts
@@ -1,0 +1,5 @@
+export const WEBMCP_INVALID_INPUT_PREFIX = "Invalid input for ";
+
+export function invalidInput(tool: string, message: string): string {
+  return `${WEBMCP_INVALID_INPUT_PREFIX}${tool}: ${message}.`;
+}
